% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{fixed}
\alias{fixed}
\title{Compare literal bytes in the string}
\usage{
fixed(pattern, ignore_case = FALSE)
}
\arguments{
\item{pattern}{Pattern to modify behavior.}

\item{ignore_case}{Should case differences be ignored in the match?}
}
\value{
An integer vector.
}
\description{
Dependency-free drop-in alternative for \code{stringr::fixed()}.
This is very fast,
but not usually what you want for non-ASCII character sets.
}
