% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_count.R
\name{str_count}
\alias{str_count}
\title{Count the number of matches in a string}
\source{
Adapted from the \href{https://stringr.tidyverse.org/}{stringr} package.
}
\usage{
str_count(string, pattern = "")
}
\arguments{
\item{string}{Input vector.
Either a character vector, or something coercible to one.}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression,
as described in \link[base:regex]{base::regex}.
Control options with \code{\link[=regex]{regex()}}.

Match a fixed string (i.e. by comparing only bytes), using \code{\link[=fixed]{fixed()}}.
This is fast, but approximate.}
}
\value{
An integer vector.
}
\description{
Dependency-free drop-in alternative for \code{stringr::str_count()}.
}
