% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_dup.R
\name{str_dup}
\alias{str_dup}
\title{Duplicate and concatenate strings within a character vector}
\source{
Adapted from the \href{https://stringr.tidyverse.org/}{stringr} package.
}
\usage{
str_dup(string, times)
}
\arguments{
\item{string}{Input character vector.}

\item{times}{Number of times to duplicate each string.}
}
\value{
A character vector.
}
\description{
Dependency-free drop-in alternative for \code{stringr::str_dup()}.
}
