% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_length.R
\name{str_length}
\alias{str_length}
\title{Compute the length of a string}
\source{
Adapted from the \href{https://stringr.tidyverse.org/}{stringr} package.
}
\usage{
str_length(string)
}
\arguments{
\item{string}{Input vector.
Either a character vector, or something coercible to one.}
}
\value{
A numeric vector the same length as string.
}
\description{
Dependency-free drop-in alternative for \code{stringr::str_length()}.
}
