% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_trim.R
\name{str_squish}
\alias{str_squish}
\title{Remove whitespace}
\source{
Adapted from the \href{https://stringr.tidyverse.org/}{stringr} package.
}
\usage{
str_squish(string)
}
\arguments{
\item{string}{Input vector.
Either a character vector, or something coercible to one.}
}
\value{
A character vector the same length as \code{string}.
}
\description{
Dependency-free drop-in alternative for \code{stringr::str_squish()}.
}
