% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_trim.R
\name{str_trim}
\alias{str_trim}
\title{Remove whitespace}
\source{
Adapted from the \href{https://stringr.tidyverse.org/}{stringr} package.
}
\usage{
str_trim(string, side = c("both", "left", "right"))
}
\arguments{
\item{string}{Input vector.
Either a character vector, or something coercible to one.}

\item{side}{Side on which to remove whitespace:
\code{"left"}, \code{"right"}, or \code{"both"}, the default.}
}
\value{
A character vector the same length as \code{string}.
}
\description{
Dependency-free drop-in alternative for \code{stringr::str_trim()}.
}
