% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_length.R
\name{str_width}
\alias{str_width}
\title{Compute the width of a string}
\source{
Adapted from the \href{https://stringr.tidyverse.org/}{stringr} package.
}
\usage{
str_width(string)
}
\arguments{
\item{string}{Input vector.
Either a character vector, or something coercible to one.}
}
\value{
A numeric vector the same length as string.
}
\description{
Dependency-free drop-in alternative for \code{stringr::str_width()}.
Results for non-ASCII characters may be inaccurate in R < 4.0.
}
