% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringstatic-package.R
\docType{package}
\name{stringstatic-package}
\alias{stringstatic}
\alias{stringstatic-package}
\title{stringstatic: Dependency-Free String Operations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides drop-in replacements for functions from the 'stringr' package, with the same user interface. These functions have no external dependencies and can be copied directly into your package code using the 'staticimports' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rossellhayes/stringstatic}
  \item Report bugs at \url{https://github.com/rossellhayes/stringstatic/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) (str_pad() and str_split() functions) [contributor]
  \item Hadley Wickham \email{hadley@rstudio.com} (stringr package) [contributor, copyright holder]
  \item RStudio (stringr package) [copyright holder]
}

}
\keyword{internal}
