% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{regex}
\alias{regex}
\title{Control regex matching behavior}
\source{
Adapted from the \href{https://stringr.tidyverse.org/}{stringr} package.
}
\usage{
regex(
  pattern,
  ignore_case = FALSE,
  multiline = FALSE,
  comments = FALSE,
  dotall = FALSE
)
}
\arguments{
\item{pattern}{Pattern to modify behavior.}

\item{ignore_case}{Should case differences be ignored in the match?}

\item{multiline}{If \code{TRUE}, \code{$} and \code{^} match the beginning and end of each line.
If \code{FALSE}, the default, only match the start and end of the input.}

\item{comments}{If \code{TRUE}, white space and comments beginning with \verb{#} are ignored.
Escape literal spaces with \verb{\\\\}.}

\item{dotall}{If \code{TRUE}, \code{.} will also match line terminators.}
}
\value{
An integer vector.
}
\description{
Dependency-free drop-in alternative for \code{stringr::regex()}.
}
